requirements.txt
tensorflow==2.9.2
numpy==1.21.6
keras==2.9.0
matplotlib==3.2.2
Pillow==7.1.2
opencv-python==4.6.0.66
ipykernel==5.3.4
scikit-learn==1.0.2
jupyter
